moved Main to new file structure
