--- src/main/kotlin/hwr/oop/Game.kt
@@ -1,10 +1,0 @@
-package hwr.oop
-
-class Game (
-    val id: Int,
-) {
-
-    init {
-        GameStorage.saveGame(this)
-    }
-}
