--- src/main/kotlin/hwr/oop/board/Board.kt
@@ -1,6 +1,0 @@
-package hwr.oop.board
-
-
-class Board () {
-    var fen: String = ""
-}
