--- src/main/kotlin/hwr/oop/board/Location.kt
@@ -1,3 +1,0 @@
-package hwr.oop.board
-
-class Location(val x: Char, val y: Int) {}
