--- src/main/kotlin/hwr/oop/pieces/Bishop.kt
@@ -1,6 +1,0 @@
-package hwr.oop.pieces
-
-import hwr.oop.board.Location
-
-class Bishop: Piece(Location('E', 4), 3) {
-}
