--- src/main/kotlin/hwr/oop/pieces/King.kt
@@ -1,6 +1,0 @@
-package hwr.oop.pieces
-
-import hwr.oop.board.Location
-
-class King: Piece(Location('E', 4), 10) {
-}
