--- src/main/kotlin/hwr/oop/pieces/Knight.kt
@@ -1,6 +1,0 @@
-package hwr.oop.pieces
-
-import hwr.oop.board.Location
-
-class Knight: Piece(Location('E', 4), 3) {
-}
