--- src/main/kotlin/hwr/oop/pieces/Pawn.kt
@@ -1,6 +1,0 @@
-package hwr.oop.pieces
-
-import hwr.oop.board.Location
-
-class Pawn: Piece(Location('E', 4), 1) {
-}
