--- src/main/kotlin/hwr/oop/pieces/Piece.kt
@@ -1,6 +1,0 @@
-package hwr.oop.pieces
-
-import hwr.oop.board.Location
-
-open class Piece(val location: Location, val value: Int) {
-}
