--- src/main/kotlin/hwr/oop/pieces/Queen.kt
@@ -1,6 +1,0 @@
-package hwr.oop.pieces
-
-import hwr.oop.board.Location
-
-class Queen: Piece(Location('E', 4), 9) {
-}
