--- src/main/kotlin/hwr/oop/pieces/Rook.kt
@@ -1,6 +1,0 @@
-package hwr.oop.pieces
-
-import hwr.oop.board.Location
-
-class Rook: Piece(Location('E', 4), 5) {
-}
