--- src/main/kotlin/hwr/oop/player/Player.kt
@@ -1,7 +1,0 @@
-package hwr.oop.player
-
-class Player (
-    val id: Int
-) {
-   // var points: Int = 0
-}
