--- src/test/kotlin/hwr/oop/MoveTest.kt
@@ -1,6 +1,0 @@
-package hwr.oop
-
-import io.kotest.core.spec.style.AnnotationSpec
-
-class MoveTest : AnnotationSpec() {
-}
