--- src/test/kotlin/hwr/oop/board/BoardTest.kt
@@ -1,6 +1,0 @@
-package hwr.oop.board
-
-import io.kotest.core.spec.style.AnnotationSpec
-
-class BoardTest : AnnotationSpec() {
-}
