--- src/test/kotlin/hwr/oop/board/LocationTest.kt
@@ -1,6 +1,0 @@
-package hwr.oop.board
-
-import io.kotest.core.spec.style.AnnotationSpec
-
-class LocationTest : AnnotationSpec() {
-}
