--- src/test/kotlin/hwr/oop/pieces/KnightTest.kt
@@ -1,6 +1,0 @@
-package hwr.oop.pieces
-
-import io.kotest.core.spec.style.AnnotationSpec
-
-class KnightTest : AnnotationSpec() {
-}
