--- src/test/kotlin/hwr/oop/pieces/QueenTest.kt
@@ -1,6 +1,0 @@
-package hwr.oop.pieces
-
-import io.kotest.core.spec.style.AnnotationSpec
-
-class QueenTest : AnnotationSpec() {
-}
