--- src/test/kotlin/hwr/oop/player/PlayerTest.kt
@@ -1,6 +1,0 @@
-package hwr.oop.player
-
-import io.kotest.core.spec.style.AnnotationSpec
-
-class PlayerTest : AnnotationSpec() {
-}
